Override on_punch for poster node definition
The signs_api `register_sign`'s default `on_punch` sets a sign's formspec to the
sign editor. If we protect a poster and then punch/break it, subsequent attempts
to view the poster will briefly cause the the sign editor to display before the
actual poster formspec. Small, but annoying.